fixed operations using stack

+++ compiler/src/analyzer/mod.rs
@@ -188,8 +188,8 @@ impl Analyzer<'_> {
 
                 if match op {
                     Op::And | Op::Or => type1 != "bool" || type2 != "bool",
-                    _ => type1 != "i32" || type2 != "i32",
-                } && (type1 != "ptr" || type2 != "i32")
+                    _ => type1 != "i8" || type2 != "i8",
+                } && (type1 != "ptr" || type2 != "i8")
                 {
                     self.errors.add(
                         ErrorKind::OperationNotImplemented {
@@ -224,10 +224,10 @@ impl Analyzer<'_> {
 
                 match name.as_str() {
                     "exit" => {
-                        if type1 != "i32" {
+                        if type1 != "i8" {
                             self.errors.add(
                                 ErrorKind::MismatchedTypes {
-                                    expected: "i32".to_string(),
+                                    expected: "i8".to_string(),
                                     found: type1,
                                 },
                                 arg.line,
@@ -270,7 +270,7 @@ impl Analyzer<'_> {
                     "null"
                 }
             }
-            ExpressionKind::IntLit(..) | ExpressionKind::Idx(..) => "i32",
+            ExpressionKind::IntLit(..) | ExpressionKind::Idx(..) => "i8",
             ExpressionKind::BoolLit(..) | ExpressionKind::Not(..) => "bool",
             ExpressionKind::StringLit(..) | ExpressionKind::ArrayLit(..) => "ptr",
             ExpressionKind::Op(_, op, _) => op.get_type(),
